Apply a more declarative API output format logic
API output formats are now driven by classes in the `*\Api\Output` namespace which implement the \Nails\Api\Interfaces\Output interface. These classes declare the Content-Type header to use as well as provide logic for actualy rendering the output. This allows for modules, or the app, to add or override output formatting logic.

Additionally, this commit introduces the `body` proeprty of ApiResponse which, if specified, will be the exact body of the response. This allows for the developer to completely customise the response if necessary.